feat: convert cjsToEsm to EcmaScript
